Backend/requirements.txt
flask
bs4
requests
firebase_admin
pycryptodome
flask-cors
base64